@@ -12,31 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.8",
-    "@angular/common": "^7.2.8",
-    "@angular/compiler": "^7.2.8",
-    "@angular/core": "^7.2.8",
-    "@angular/forms": "^7.2.8",
-    "@angular/http": "^7.2.8",
-    "@angular/platform-browser": "^7.2.8",
-    "@angular/platform-browser-dynamic": "^7.2.8",
-    "@angular/router": "^7.2.8",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "core-js": "^2.6.5",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "three": "^0.104.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "^7.3.5",
-    "@angular/compiler-cli": "^7.2.8",
-    "@angular/language-service": "^7.2.8",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "~3.3.9",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "~12.0.0",
     "@types/three": "^0.103.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
@@ -44,12 +43,12 @@
     "karma-coverage-istanbul-reporter": "~2.0.5",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.0.1",
+    "ng-packagr": "^5.2.0",
     "protractor": "~6.0.0",
     "ts-node": "~8.2.0",
     "tsickle": "^0.35.0",
     "tslib": "^1.9.3",
     "tslint": "~5.14.0",
-    "typescript": "3.2.4"
+    "typescript": "3.4.5"
   }
 }
